requirements.txt
Flask==3.1.2
psycopg2-binary==2.9.11
gunicorn==23.0.0
boto3==1.40.16
botocore==1.40.16
python-dotenv==1.1.1

# HyperDX OpenTelemetry instrumentation
hyperdx-opentelemetry

# Additional OpenTelemetry packages for comprehensive instrumentation
opentelemetry-instrumentation[otlp]
opentelemetry-instrumentation-flask
opentelemetry-instrumentation-requests
opentelemetry-instrumentation-urllib3
opentelemetry-instrumentation-boto3sqs
